Change Id from int to long

+++ ResumableFunctions.Handler/Core/Abstraction/ICallPusher.cs
@@ -4,7 +4,7 @@ namespace ResumableFunctions.Handler.Core.Abstraction
 {
     public interface ICallPusher
     {
-        Task<int> PushCall(PushedCall pushedCall);
-        Task<int> PushExternalCall(PushedCall pushedCall, string serviceName);
+        Task<long> PushCall(PushedCall pushedCall);
+        Task<long> PushExternalCall(PushedCall pushedCall, string serviceName);
     }
 }
