+++ ResumableFunctions.Handler/DataAccess/Abstraction/IWaitTemplatesRepo.cs
@@ -6,9 +6,9 @@ namespace ResumableFunctions.Handler.DataAccess.Abstraction
     public interface IWaitTemplatesRepo
     {
         Task<WaitTemplate> AddNewTemplate(
-            WaitExpressionsHash hashResult, object currentFunctionInstance, int funcId, int groupId, int methodId);
-        Task<WaitTemplate> CheckTemplateExist(byte[] hash, int funcId, int groupId);
-        Task<List<WaitTemplate>> GetWaitTemplates(int methodGroupId);
-        Task<WaitTemplate> GetById(int templateId);
+            WaitExpressionsHash hashResult, object currentFunctionInstance, long funcId, long groupId, long methodId);
+        Task<WaitTemplate> CheckTemplateExist(byte[] hash, long funcId, long groupId);
+        Task<List<WaitTemplate>> GetWaitTemplates(long methodGroupId);
+        Task<WaitTemplate> GetById(long templateId);
     }
 }
