+++ ResumableFunctions.Handler/DataAccess/WaitsRepo.cs
@@ -70,7 +70,7 @@ internal partial class WaitsRepo : IWaitsRepo
         }
     }
 
-    public async Task<List<int>> GetMatchedFunctionsForCall(int pushedCallId, string methodUrn)
+    public async Task<List<long>> GetMatchedFunctionsForCall(long pushedCallId, string methodUrn)
     {
         try
         {
@@ -166,7 +166,7 @@ internal partial class WaitsRepo : IWaitsRepo
         }
     }
 
-    private async Task<int> GetMethodGroupId(string methodUrn)
+    private async Task<long> GetMethodGroupId(string methodUrn)
     {
         var methodGroup =
            await _context
@@ -181,7 +181,7 @@ internal partial class WaitsRepo : IWaitsRepo
         throw new Exception(error);
     }
 
-    public async Task RemoveFirstWaitIfExist(int methodIdentifierId)
+    public async Task RemoveFirstWaitIfExist(long methodIdentifierId)
     {
         try
         {
@@ -218,11 +218,11 @@ internal partial class WaitsRepo : IWaitsRepo
     }
 
 
-    public async Task CancelSubWaits(int parentId, int pushedCallId)
+    public async Task CancelSubWaits(long parentId, long pushedCallId)
     {
         await CancelWaits(parentId);
 
-        async Task CancelWaits(int pId)
+        async Task CancelWaits(long pId)
         {
             var waits = await _context
                 .Waits
@@ -238,7 +238,7 @@ internal partial class WaitsRepo : IWaitsRepo
         }
     }
 
-    private void CancelWait(Wait wait, int pushedCallId)
+    private void CancelWait(Wait wait, long pushedCallId)
     {
         wait.LoadUnmappedProps();
         wait.Cancel();
@@ -264,14 +264,14 @@ internal partial class WaitsRepo : IWaitsRepo
     }
 
 
-    public async Task CancelOpenedWaitsForState(int stateId)
+    public async Task CancelOpenedWaitsForState(long stateId)
     {
         await _context.Waits
               .Where(x => x.FunctionStateId == stateId && x.Status == WaitStatus.Waiting)
               .ExecuteUpdateAsync(x => x.SetProperty(wait => wait.Status, status => WaitStatus.Canceled));
     }
 
-    public async Task CancelFunctionWaits(int requestedByFunctionId, int functionStateId)
+    public async Task CancelFunctionWaits(long requestedByFunctionId, long functionStateId)
     {
         var functionInstanceWaits =
             await _context.Waits
