+++ ResumableFunctions.Handler/UiService/InOuts/ServiceStatistics.cs
@@ -1,4 +1,4 @@
 ﻿namespace ResumableFunctions.Handler.UiService.InOuts
 {
-    public record ServiceStatistics(int Id, string ServiceName, int ErrorCounter, int FunctionsCount, int MethodsCount);
+    public record ServiceStatistics(long Id, string ServiceName, int ErrorCounter, int FunctionsCount, int MethodsCount);
 }
